Change i32 to usize and other minor adjustments

+++ bezier-rs/lib/src/consts.rs
@@ -14,9 +14,9 @@ pub const SCALABLE_CURVE_MAX_ENDPOINT_NORMAL_ANGLE: f64 = std::f64::consts::PI /
 /// Default `t` value used for the `curve_through_points` functions.
 pub const DEFAULT_T_VALUE: f64 = 0.5;
 /// Default LUT step size in `compute_lookup_table` function.
-pub const DEFAULT_LUT_STEP_SIZE: i32 = 10;
+pub const DEFAULT_LUT_STEP_SIZE: usize = 10;
 /// Default number of subdivisions used in `length` calculation.
-pub const DEFAULT_LENGTH_SUBDIVISIONS: i32 = 1000;
+pub const DEFAULT_LENGTH_SUBDIVISIONS: usize = 1000;
 /// Default step size for `reduce` function.
 pub const DEFAULT_REDUCE_STEP_SIZE: f64 = 0.01;
